Fix bundling errors by porting rn-cli to metro
